+++ RELEASE_NOTES.md
@@ -1,4 +1,3 @@
 # Release Notes
 
-- If the image of the product is loaded on scrapping, the image will be sent with the information of the offer.
-- The URL of the product will be sent with the information of the offer.
+- Notification time is now displayed at the /start command